chore: remove liveSyncServices from appbuilder dir
Remove all LiveSync related services from appbuilder dir (and the whole dir). Cleanup interfaces and classes used only in the deleted files.
